Fixing the .gitignore to remove unnecessary step in docs
The current "Basic Configuration" page of the Laravel documentation
has a step that requires you to create the application/config/local
directory. This is an unnecessary step that we should do for the user.

My guess is that this wasn't done previously due to git's limitation
on empty directories. However by moving the git ignore file for the
application/config/local directory into local and making it ignore
all but itself (the .gitignore file), we are able to get around this
limitation.