Fix case mismatches in Bundle loader
Bundle class mismatched cases in some comparisons, thus reporting that
any bundle with a non lowercased name would load but not be reported as
existing.

Currently I'm fixing the case mismatch in the exists() function because
I'm not sure where else "names()" is called. It may be better to run
the array_map in the names() function itself.

Also fixed a small comment typo I discovered.