Removed passing a column to count in Paginate
Removed using the column passed in because if you pass a column like
this 'table.column_name AS table_column_name' it will cause the query
to throw an error as it tries to wrap 'table.column_name AS
table_column_name' so that in the query it is "SELECT
COUNT(table.column_name AS table_column_name)….." which is invalid.

As far a pagination goes I don't see this change as affecting it much
at all.