Add Str::human() - opposite of Str::slug()
To be used with Str::title() or ucfirst() to convert slugs and
url-encoded strings into human-readable text.

The second (optional) parameter contains an array of characters to
replace with spaces. The default is `array('_', '-', '+')`.
Suggestions for default values is welcome.

Also, this method currently leaves the string's capitalization as-is.
Feedback on whether it should include any/what capitalization?

Signed-off-by: Kelly Banman <kelly.banman@gmail.com>