Updated laravel/uri.php
Fixed so that URL's work as expected and those with extraneous slashes (e.g. laravel.com////about or laravel.com/about////) 404 as desired.

See this for more details of both this and a non-regex option with speed benchmark results of the two. In sum, they benchmark at very nearly the same speed, so this version was chosen because it's cleaner and more elegant. https://github.com/laravel/laravel/issues/1412#issuecomment-10278667