Removed the 'id' value from Has_Many_And_Belongs_To::$with.
If a pivot table does not have an id field, then attempts to use $this->has_many_and_belongs_to() will fail. To keep with the 'Development should be a creative experience
that you enjoy, not something that is painful.' mentality, developers should be free to decide whether or not to have specific fields, and they should be responsible for defining the columns necessary for their pivots.