Fire log event with class & function name
In Profiler, the class & function name are not shown in log. In many cases, this would make debugging a lot easier. With this patch, the class and function name are also shown in Profiler. timestamp is less usefull in Profiler.

This is a very simple patch, but a better approach would be to add the class and function as parameters when firing, so you can decide somewhere else whether or not to show class & function in Profiler log, I suggest an optional extra config setting for this.

But that's a change "to big for me at the moment" :-)