Correct a bug in the exists function.
Update to correct a bug when using the exists function within a Model object. 
Example: when using User::where('id', '>', '0')->exists() it was returning a Builder object. With this change we can know get the boolean value as expected. 
The toSql function was also added to the passthru array in order to permit verify the SQL quereis builded for the Eloquent objects.