Prevent endless loop in .htaccess slash removal
Pointing to a folder in the URL resulted in an endless loop. Directory names automatically redirect to the same name with a trailing slash, and the rule removed the slash. With the conditions added, it won't touch to directory names.

The rule for the front controller may also be modified to include directory names, so we could have localized resources like '/fr/images/welcome.jpg' and also use localized content like 'mysite.com/fr', without the need to add something like 'mysite.com/fr/home'.