Update .htaccess
I'm not sure exactly how this impacts the entire project, but for a note on why I'm recommending it, see [this issue](https://github.com/laravel/framework/issues/5344) from the laravel/framework repo.

It doesn't seem right to force the redirect unless it's necessary. That should be left up to the developers to do if it's a visual only kind of thing, hence the suggestion.