remove google fonts
With laravel popular in China, more and more developers to join the laravel arms, but now there is a problem, most developers when installing Laravel half-day untapped home, and later found because Home Use the google font, it is proposed to remove this dependence, leaving the majority of Laravel fans a good impression. Thank you!