This is a -3 compatible encryption comment
I tried to improve the comment according to the -3 standards. According to what I have read regarding Mcrypt, the randomness and "non humaness" determines the security, not only the length. As previously discussed in similar issues, the key MUST be set to a correct length starting with PHP 5.6 and should be done before using any of Laravel's encryption features.