Fix PSR-2 standard on User model
According to PSR-2 standard, point 4.1 about extends and implements,
"Lists of implements MAY be split across multiple lines, where each
subsequent line is indented once. When doing so, the first item in the
list MUST be on the next line, and there MUST be only one interface per
line.", so User model implements are wrong.

This change should also be made in Laravel documentation pages.