Differentiate `exists` validation rule message.
The `in`/`not_in` rules had the exact same failure message as `exists` when the two seem to have quite different uses. Obviously the two would almost never be used on the same field, but it still made detection (human or programmatic) of different validation problems more complicated than necessary.