Call `shouldUse` on Auth when using the Authorize middleware
I imagine most people would expect to just be able to use `Auth::user()` or inject Guard and be able to just do `->user()` in their app if the user has been piped through this middleware as something other than the default guard. But they won't be able to without this as they won't be authed on that specific guard so they will instead receive `null` when expecting a user instance (considering they've been 'authenticated') right?.

This is an issue that annoys me when developing a simple front-end and api in the same app. Is there any particular reason this isn't already the case? I completely expect this behaviour... I'm 96% sure it's not just me.