csrfToken is auto added to window.Laravel
In `/resources/assets/js/boostrap.js` there is this header `'X-CSRF-TOKEN': window.Laravel.csrfToken,` added to the default axios headers. You get an error out of a fresh install of Laravel because the window.Laravel.csrfToken is not being set anywhere by default. You have to know about the csrf token and that Laravel expects you to put it in your html manually if you want to use the new Laravel Mix functionality and do things like `npm run dev`.