Update default homepage
Change URL for logging in and registering on default homepage. Right now Laravel checks if there is a login route and then it sends the person to /login even if their login route is somewhere else, such as at /admin/login (e.g. they are using Backpack for Laravel). Therefore, instead this will check the exact URL for the login route.