Update app.php to respect PHP date.timezone
This change updates `app.php` to use the following line instead of setting the timezone to UTC by default.

`'timezone' => env('APP_TIMEZONE', date_default_timezone_get())`

Fix for: https://github.com/laravel/framework/issues/23743

The end result here is the same. Laravel would respect the server's timezone setting, but if it was not set, the default is UTC anyway.