Use Mailhog Settings
MailTrap is a paid **external** service (with a optional free tier).  Mailhog is a free **local** service, and comes preinstalled on Homestead machines.

There may be additional things that MailTrap offers that Mailhog does not, but I think switching from an external to local service (for the default) is better.

Also, I'm not sure we should give Mailtrap preference, since there's lots of external mail catching services, and people are still free to use what they'd like.