Prepare app.js for Mix 4.0 (and future)
Since Mix 4.0 is still beta but is based on Babel7, it will now need to behave like `import`, which requires us to mention `default` to know what is `require`d.

Not sure if this is required right now, but it will be required once the switch is made. AFAIK, this is a non-breaking change for people who don't upgrade.