default to using app.timezone for mysql
MySQL defaults to using the *local system* timezone for connections, which may not be the same as the config value for `app.timezone`. During daylight savings this can result in Laravel saving invalid TIMESTAMP dates once a year for 1 hour, because MySQL internally converts the connection timezone to UTC when it saves the TIMESTAMP field as a unix timestamp.