Remove base controller from skeleton
This idea comes from the following blog post: https://murze.be/simplifying-controllers#controllers-dont-need-to-extend-anything

Basically controllers don't need to extend anything and perfectly can be used stand-alone. If you want to use some of the functionality provided by the traits used in the base controller you can simply require them on those controllers. It's one less file for the skeleton and doesn't breaks anything atm. Updated the Auth controllers with the base route controller class since they don't need any of the traits.

People are free to add this base controller if they want in their own projects.

This will need some updates to the docs as well which I'll work on if the PR gets merged.