Add npm-run-all to simplify NPM scripts
Including the full command for Laravel Mix can make the NPM scripts a
bit difficult to comprehend especially when there are a few instances of
it. `npm-run-all` provides a couple of useful commands to make scripts
simpler to write.

In this case, I’ve added a new script called `mix` which will serve as
the base command which we can use for our other tasks.

`run-s` runs a list of commands (with support for glob-patterns)
sequentially.

Each script now clearly indicates which environment its running for, as
well, which webpack options are set.

Note: we still do need to use escaped double quotes to pass flags and
arguments to the scripts since apparently there are issues
cross-platform if you just use unescaped single quotes.