Flash status message into verified view session
When a new user receives the "Verify Email Address" email and clicks on the included link they should be redirected to the default HOME route with a message injected into the session to display to the user for the HOME view to render. This lets the user confirm the Verify Email Address link worked correctly.