Fix mailhog host.
This fixes the mailhog host on clean installs. Otherwise, we get the error `Connection could not be established with host mailhog`.

This value should be set to `localhost` as per the homestead mailhog config docs: https://laravel.com/docs/8.x/homestead#configuring-mailhog

Setting this value to `localhost` resolves the issue out of the box. 

P.S. mailhog rocks :)