Stop Env Commits
This is not a laravel issue per se, but there are a portion of users who
change the .env.example file with real data and then commmit it, or even
just commit the generated .env file.

This change adds a hook in to check the commits to see if they contain
an .env file and gives users a way to commit either way (ie they
do know what they are doing and want to skip over the message, or they
were not aware of what they were doing and a way to recover from that
(without losing their changes).

There are a few bots that are actively scanning for these files as
laravel has a large user base and people commit to this file. AWS
Credentials and mail credentials are most common.

This may not be the best solution, but happy to have a discussion on how
to prevent users of laravel making these mistakes which can be costly.