Removed MAIL_ENCRYPTION and MAIL_FROM_ADDRESS env vars
Variables set in the .env as null or empty do not use the default value when calling env(). In some cases this produces errors, for example when sending an email. Symfony Mailer needs a From/Sender header and because MAIL_FROM_ADDRESS value is null in the .env, the default set in the config/mail.php is not used. This also occurs for the MAIL_ENCRYPTION.