Add .htaccess for Improved Hosting Compatibility
This commit adds a new .htaccess file in the root of the Laravel project. It ensures that the application functions correctly in hosting environments without direct access to the vhost configuration. For Nginx, the configuration should be root /var/www/html/public;, and for Apache, it should be DocumentRoot /var/www/laravel/public.

With this .htaccess file, installation on hosting services that do not provide full control over the directory structure is simplified, redirecting all requests to the public directory.